pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Efim Kashapov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
black = "^24.10.0"
isort = "^5.13.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
